【PR】: sync ge_dev to master 20260326

Created-by: peiyang
Commit-by: cann-robot;liyanrong_9;wanghuajie;wuzhengzhen;coconut_tree;j18351891940;isaacxu;chenyukai;gaoshuer913;wqtshg_wt;kobemini;lining23666;yelongjian;stevenaw0;zhang_shengjie;yuht9;xchu42;yangzw653;peiyang;ApeiriaNode_Booker;jsong27;li-yuanjie-da;jikai-tyler;chengyutao3;tangqunzhang;GRJ_XIDUO
Merged-by: peiyang
Description: 

# Pull Request

## 描述
本PR指在将develop分支的代码同步合入master分支，包括以下特性：
1、compare接口，输入为S64时，regtensor改用regTraitNumTwo
2、编译选项整改
3、SqueezeV3/UnSqueezeV3/FlattenV2 Infer Symbol Shape
4、RedundantControlEdgeRemovePass支持多ctrl输入场景
5、readme文件新增GE小助手二维码
6、删除冗余函数&解决动态shape无法判断repeat的问题
7、launch切换aclrt接口

## 变更类型
请选择本次引入的变更类型：
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新功能
- [ ] 💄 代码风格更新（格式化，局部变量）
- [ ] ♻️ 重构（既不修复错误也不增加功能的代码变动）
- [ ] 📦 构建过程或辅助工具的变动
- [ ] 📝 文档内容更新

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在当前页面的右侧'关联Issue'部分添加相应Issue链接，并勾选'合并后关闭已关联的 Issue'选项。 -->


## 如何测试
描述测试此变更的步骤和前提条件：
1.
2.

## 核对清单
<!--  [x] 表示选中 -->
- [ ] 我的代码遵循了项目的代码风格
- [ ] 我已对代码进行了自测
- [ ] 我已更新了相关的文档
- [ ] 我在标题中使用了合适的类型标签（如：`feat:`, `fix:`）
- [ ] 我已经详细阅读了贡献指南（CONTRIBUTING.md），并遵守了其中的所有规定，包括但不限于commit message的格式、无效commit的合并等

## 其他信息
在此添加任何其他关于本次 PR 的说明。




See merge request: cann/ge!1508

+++ tests/autofuse/v35/ut/ascendc/api_regbase/test_where.cpp
@@ -207,7 +207,7 @@ class TestRegbaseApiWhereUT : public testing::Test {
     GmToUb(l_y, param.y, param.y_stride * param.m);
 
     if constexpr (dim == 1) {
-        const uint16_t output_dims[dim] = {param.size};
+        const uint16_t output_dims[dim] = {(uint16_t)param.size};
         const uint16_t output_stride[dim] = {1};
         const uint16_t mask_stride[dim] = {1};
         const uint16_t input_stride[dim] = {1};
@@ -222,10 +222,10 @@ class TestRegbaseApiWhereUT : public testing::Test {
           WhereExtend<false, false, 1, T, T, T>(l_y, l_x1, l_x2, l_x3, output_dims, output_stride, mask_stride, input_stride);
         }
     } else if constexpr (dim == 2) {
-        const uint16_t output_dims[dim] = {param.m, param.size};
-        const uint16_t output_stride[dim] = {param.y_stride, 1};
-        const uint16_t mask_stride[dim] = {param.x1_stride, 1};
-        const uint16_t input_stride[dim] = {param.x2_stride, 1};
+        const uint16_t output_dims[dim] = {(uint16_t)param.m, (uint16_t)param.size};
+        const uint16_t output_stride[dim] = {(uint16_t)param.y_stride, 1};
+        const uint16_t mask_stride[dim] = {(uint16_t)param.x1_stride, 1};
+        const uint16_t input_stride[dim] = {(uint16_t)param.x2_stride, 1};
         
         if (param.x2_bcast && param.x3_bcast) {
           WhereExtend<true, true, 2, T, T, T>(l_y, l_x1, l_x2_, l_x3_, output_dims, output_stride, mask_stride, input_stride);
